requirements.txt
langchain
langchain-google-genai
langchain-chroma
chromadb
python-dotenv
langchain-text-splitters